% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{add_dosing_events}
\alias{add_dosing_events}
\title{Adds dosing events for a model from a NONMEM dataframe
Works for bolus doses and infusions.
Function does not modify the input event table.}
\usage{
add_dosing_events(event_table, model, df)
}
\arguments{
\item{event_table}{The event table that should added to}

\item{model}{The target model}

\item{df}{The NONMEM dataframe}
}
\value{
A modified event table with added dosing events
}
\description{
Adds dosing events for a model from a NONMEM dataframe
Works for bolus doses and infusions.
Function does not modify the input event table.
}
