% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{create_init}
\alias{create_init}
\title{Creates a init vector for RxODE from the model compartment list}
\usage{
create_init(model)
}
\arguments{
\item{model}{the RxODE model}
}
\value{
Init vector initialized with 0s
}
\description{
Creates a init vector for RxODE from the model compartment list
}
