% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{create_obj_fn}
\alias{create_obj_fn}
\title{Creates an MAP objective function}
\usage{
create_obj_fn(
  individual_df,
  model,
  event_table,
  inits,
  THETA,
  I_OMEGA,
  SIGMA,
  iCov = NULL,
  error_model = c("proportional", "additive"),
  ...
)
}
\arguments{
\item{individual_df}{NONMEM dataframe with individual measurements}

\item{model}{the RxODE model}

\item{event_table}{the RxODE event table}

\item{inits}{a vector of initial values of the state variables (e.g., amounts in each compartment),
and the order in this vector must be the same as the state variables (e.g., PK/PD compartments)#'}

\item{THETA}{Typical values THETA}

\item{I_OMEGA}{Inverse of model OMEGA as matrix}

\item{SIGMA}{residual variance SIGMA of the model}

\item{iCov}{dataframe of individual time-independent covariates or NULL if not used}

\item{error_model}{The residual error model/function}

\item{...}{additonal parameters for RxODE 'solve' function}
}
\value{
the MAP objective function
}
\description{
Creates an MAP objective function
}
