% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{map_estimate}
\alias{map_estimate}
\title{MAP estimation for model ETAs for an individual}
\usage{
map_estimate(
  individual_df,
  model,
  event_table,
  inits,
  THETA,
  OMEGA,
  SIGMA,
  iCov = NULL,
  error_model = c("proportional", "additive"),
  fit_options = list(method = "L-BFGS-B"),
  ...
)
}
\arguments{
\item{individual_df}{NONMEM dataframe with measured data of the individual for the estimation}

\item{model}{RxODE model}

\item{event_table}{RxODE event table}

\item{inits}{a vector of initial values of the state variables (e.g., amounts in each compartment),
and the order in this vector must be the same as the state variables (e.g., PK/PD compartments)#'}

\item{THETA}{values of structural THETA parameters of the model.}

\item{OMEGA}{variance of intersubject errors. Must be a matrix or vector.}

\item{SIGMA}{variance}

\item{iCov}{dataframe of individual time-independent covariates or NULL if not used}

\item{error_model}{The residual error model/function}

\item{fit_options}{Options for the optim fitting routine}

\item{...}{additonal parameters for RxODE 'solve' function}
}
\value{
List with optimized parameters, hessian, information about convergence and standard errors of the fit.
}
\description{
MAP estimation for model ETAs for an individual
}
