% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{map_obj_fn}
\alias{map_obj_fn}
\title{Objective function of the empicical bayes estimate (EBE)}
\usage{
map_obj_fn(
  ETA,
  I_OMEGA,
  C_i,
  C_m,
  SIGMA,
  error_model = c("proportional", "additive")
)
}
\arguments{
\item{ETA}{ETAs of PK/PD-Parameter of the individual as row vector}

\item{I_OMEGA}{Inverve of the IIV of the PK-Parameter (interindividual variance)}

\item{C_i}{observed values of the individual}

\item{C_m}{simulated values for the individual}

\item{SIGMA}{Residual variance of the model}

\item{error_model}{The residual error model/function}
}
\value{
Objective function value
}
\description{
Objective function of the empicical bayes estimate (EBE)
}
