% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{obs_times}
\alias{obs_times}
\title{Extracts observed times from a NONMEM dataframe}
\usage{
obs_times(df, include_events = FALSE)
}
\arguments{
\item{df}{the NONMEM dataframe}

\item{include_events}{TRUE if event times should be included}
}
\value{
A vector of times with unique times. Result will be sorted.
}
\description{
Extracts observed times from a NONMEM dataframe
}
