% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{obs_values}
\alias{obs_values}
\title{Extracts observed values from a NONMEM dataframe}
\usage{
obs_values(df, include_time = FALSE)
}
\arguments{
\item{df}{the NONMEM dataframe}

\item{include_time}{TRUE, if the time column should be extracted as well, else FALSE}
}
\value{
A dataframe with observed values
}
\description{
Extracts observed values from a NONMEM dataframe
}
