% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{simulate_model}
\alias{simulate_model}
\title{Simulates an RxODE model}
\usage{
simulate_model(
  model,
  event_table,
  inits,
  THETA,
  sim_times,
  ETA = NULL,
  iCov = NULL,
  ...
)
}
\arguments{
\item{model}{the RxODE model}

\item{event_table}{the RxODE event table}

\item{inits}{a vector of initial values of the state variables (e.g., amounts in each compartment),
and the order in this vector must be the same as the state variables (e.g., PK/PD compartments)}

\item{THETA}{model THETAS}

\item{sim_times}{simulation times}

\item{iCov}{dataframe of individual time-independent covariates or NULL if not used}

\item{...}{additonal parameters for RxODE 'solve' function}

\item{eta}{model ETAS or NULL if only typical values should be simulated}
}
\value{
RxODE simulation result
}
\description{
Simulates an RxODE model
}
