% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{split_by_id}
\alias{split_by_id}
\title{Splits dataframe into list of dataframes by id column 'ID'}
\usage{
split_by_id(df)
}
\arguments{
\item{df}{THe data.frame}
}
\value{
A named list of data.frames with names = ID
}
\description{
Splits dataframe into list of dataframes by id column 'ID'
}
